services/codexathenae/backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
pydantic>=2.7
motor>=3.4
pymongo>=4.7
loguru>=0.7
python-dotenv>=1.0
orjson>=3.10
pytest>=8.2
pytest-asyncio>=0.23
httpx>=0.27
